pyproject.toml
[project]
name = "opendoctorate-ex-public"
version = "2.6.21"
description = ""
authors = [
    {name = "MVXTP",email = "github@users.noreply.github.com"},
    {name = "Cryin",email = "github@users.noreply.github.com"},
    {name = "Autumn174",email = "github@users.noreply.github.com"}
]
license = {text = "GPL-3.0"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pycryptodome (>=3.23.0,<4.0.0)",
    "msgspec (>=0.19.0,<0.20.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
